test_echo now does not need an operator

+++ rbl/unittest.h
@@ -130,7 +130,7 @@ void UTRecordAssertResult(const char* fn, const char* file, int line, const char
 #define UT_TRUE(a) do{\
 	if(!(a))  {\
 	    char* msg; \
-		int _x_ = asprintf(&msg, BRIGHT_RED("FAILED report ") BRIGHT_CYAN("func:") BLUE(" %s") BRIGHT_CYAN(" file:") BLUE(" %s ") BRIGHT_CYAN("line:") BLUE("%d \n"), __FUNCTION__, __FILE__, __LINE__,(int)(a));\
+		int _x_ = (int)asprintf(&msg, BRIGHT_RED("FAILED report ") BRIGHT_CYAN("func:") BLUE(" %s") BRIGHT_CYAN(" file:") BLUE(" %s ") BRIGHT_CYAN("line:") BLUE("%d \n"), __FUNCTION__, __FILE__, __LINE__,(int)(a));\
         UTRecordAssertResult(__FUNCTION__, __FILE__, __LINE__, msg);   \
         free(msg); \
 		return 1; \
